refactor set target code integrate openocd board
