+++ src/espIdf/setTarget/getTargets.ts
@@ -17,153 +17,14 @@
  */
 import { EOL } from "os";
 import { join } from "path";
-import { ConfigurationTarget, Progress, ProgressLocation, Uri, window, WorkspaceFolder } from "vscode";
-import { readParameter, writeParameter } from "../../idfConfiguration";
-import { Logger } from "../../logger/logger";
-import { OutputChannel } from "../../logger/outputChannel";
-import { appendIdfAndToolsToPath, setCCppPropertiesJsonCompilerPath, spawn } from "../../utils";
-import { ConfserverProcess } from "../menuconfig/confServerProcess";
-import { getBoards, getOpenOcdScripts } from "../openOcd/boardConfiguration";
+import { Uri } from "vscode";
+import { readParameter } from "../../idfConfiguration";
+import { appendIdfAndToolsToPath, spawn } from "../../utils";
 
 export interface IdfTarget {
   label: string;
   isPreview: boolean;
   target: string;
-}
-
-export async function getSupportedTargets(placeHolderMsg: string) {
-  const configurationTarget = ConfigurationTarget.WorkspaceFolder;
-  let workspaceFolder = await window.showWorkspaceFolderPick({
-    placeHolder: `Pick Workspace Folder to which settings should be applied`,
-  });
-  if (!workspaceFolder) {
-    return;
-  }
-  const targetsFromIdf = await getTargetsFromEspIdf(workspaceFolder.uri);
-  const selectedTarget = await window.showQuickPick(targetsFromIdf);
-  if (!selectedTarget) {
-    return;
-  }
-  if (selectedTarget.target === "custom") {
-    const currentValue = readParameter(
-      "idf.customAdapterTargetName",
-      workspaceFolder.uri
-    ) as string;
-    const customIdfTarget = await window.showInputBox({
-      placeHolder: placeHolderMsg,
-      value: currentValue,
-    });
-    if (!customIdfTarget) {
-      return;
-    }
-    await writeParameter(
-      "idf.adapterTargetName",
-      selectedTarget.target,
-      configurationTarget,
-      workspaceFolder.uri
-    );
-    await writeParameter(
-      "idf.customAdapterTargetName",
-      customIdfTarget,
-      configurationTarget,
-      workspaceFolder.uri
-    );
-    return Logger.infoNotify(
-      `IDF_TARGET has been set to custom. Remember to set the configuration files for OpenOCD`
-    );
-  }
-  await writeParameter(
-    "idf.adapterTargetName",
-    selectedTarget.target,
-    configurationTarget,
-    workspaceFolder.uri
-  );
-  const openOcdScriptsPath = getOpenOcdScripts(workspaceFolder.uri);
-  const boards = await getBoards(openOcdScriptsPath);
-
-  const boardsForTarget = boards.filter((b) => b.target === selectedTarget.target);
-  const choices = boards.map((b) => {
-    return {
-      description: `${b.description} (${b.configFiles})`,
-      label: b.name,
-      target: b,
-    };
-  });
-  const selectedBoard = await window.showQuickPick(choices, {
-    placeHolder: "Enter OpenOCD Configuration File Paths list",
-  });
-  if (!selectedBoard) {
-    return;
-  }
-  await writeParameter(
-    "idf.openOcdConfigs",
-    selectedBoard.target.configFiles,
-    configurationTarget,
-    workspaceFolder.uri
-  );
-
-  await window.withProgress(
-    {
-      cancellable: false,
-      location: ProgressLocation.Notification,
-      title: "ESP-IDF: Setting device target...",
-    },
-    async (
-      progress: Progress<{ message: string; increment: number }>
-    ) => {
-      try {
-        if (ConfserverProcess.exists()) {
-          ConfserverProcess.dispose();
-        }
-        const idfPathDir = readParameter(
-          "idf.espIdfPath",
-          workspaceFolder.uri
-        );
-        const idfPy = join(idfPathDir, "tools", "idf.py");
-        const modifiedEnv = appendIdfAndToolsToPath(
-          workspaceFolder.uri
-        );
-        modifiedEnv.IDF_TARGET = undefined;
-        const enableCCache = readParameter(
-          "idf.enableCCache",
-          workspaceFolder.uri
-        ) as boolean;
-        const setTargetArgs: string[] = [idfPy];
-        if (enableCCache) {
-          modifiedEnv.IDF_CCACHE_ENABLE = "1";
-        } else {
-          modifiedEnv.IDF_CCACHE_ENABLE = undefined;
-        }
-        setTargetArgs.push("set-target", selectedTarget.target);
-        if (selectedTarget.isPreview) {
-          setTargetArgs.push("--preview");
-        }
-        const pythonBinPath = readParameter(
-          "idf.pythonBinPath",
-          workspaceFolder.uri
-        ) as string;
-        const setTargetResult = await spawn(
-          pythonBinPath,
-          setTargetArgs,
-          {
-            cwd: workspaceFolder.uri.fsPath,
-            env: modifiedEnv,
-          }
-        );
-        Logger.info(setTargetResult.toString());
-        OutputChannel.append(setTargetResult.toString());
-        setCCppPropertiesJsonCompilerPath(workspaceFolder.uri);
-      } catch (err) {
-        if (err.message && err.message.indexOf("are satisfied") > -1) {
-          Logger.info(err.message.toString());
-          OutputChannel.append(err.message.toString());
-        } else {
-          Logger.errorNotify(err, err);
-          OutputChannel.append(err);
-        }
-      }
-    }
-  );
 }
 
 export async function getTargetsFromEspIdf(workspaceFolder: Uri) {
