+++ src/extension.ts
@@ -119,6 +119,7 @@ import { WelcomePanel } from "./welcome/panel";
 import { getWelcomePageInitialValues } from "./welcome/welcomeInit";
 import { selectDfuDevice } from "./flash/dfu";
 import { getEspMatter } from "./espMatter/espMatterDownload";
+import { setIdfTarget } from "./espIdf/setTarget";
 
 // Global variables shared by commands
 let workspaceRoot: vscode.Uri;
@@ -1386,205 +1387,8 @@ export async function activate(context: vscode.ExtensionContext) {
         "extension.enterDeviceTargetMessage",
         "Enter target name (IDF_TARGET)"
       );
-      const selectedTarget = await vscode.window.showQuickPick(
-        [
-          { description: "ESP32", label: "ESP32", target: "esp32" },
-          { description: "ESP32-S2", label: "ESP32-S2", target: "esp32s2" },
-          {
-            description: "ESP32-S3 (Built-in USB JTAG)",
-            label: "ESP32-S3 (Built-in USB JTAG)",
-            target: "esp32s3",
-            type: "usb",
-          },
-          {
-            description: "ESP32-S3 (ESP-PROG JTAG)",
-            label: "ESP32-S3 (ESP-PROG JTAG)",
-            target: "esp32s3",
-            type: "prog",
-          },
-          {
-            description: "ESP32-C3 (Built-in USB JTAG)",
-            label: "ESP32-C3 (Built-in USB JTAG)",
-            target: "esp32c3",
-            type: "usb",
-          },
-          {
-            description: "ESP32-C3 (ESP-PROG JTAG)",
-            label: "ESP32-C3 (ESP-PROG JTAG)",
-            target: "esp32c3",
-            type: "prog",
-          },
-          {
-            description: "Custom target",
-            label: "Custom target",
-            target: "custom",
-          },
-        ],
-        { placeHolder: enterDeviceTargetMsg }
-      );
-      if (!selectedTarget) {
-        return;
-      }
-      const configurationTarget = vscode.ConfigurationTarget.WorkspaceFolder;
-      let workspaceFolder = await vscode.window.showWorkspaceFolderPick({
-        placeHolder: `Pick Workspace Folder to which settings should be applied`,
-      });
-      if (!workspaceFolder) {
-        return;
-      }
-      if (selectedTarget.target === "custom") {
-        const currentValue = idfConf.readParameter(
-          "idf.customAdapterTargetName",
-          workspaceFolder.uri
-        ) as string;
-        const customIdfTarget = await vscode.window.showInputBox({
-          placeHolder: enterDeviceTargetMsg,
-          value: currentValue,
-        });
-        if (!customIdfTarget) {
-          return;
-        }
-        await idfConf.writeParameter(
-          "idf.adapterTargetName",
-          selectedTarget.target,
-          configurationTarget,
-          workspaceFolder.uri
-        );
-        await idfConf.writeParameter(
-          "idf.customAdapterTargetName",
-          customIdfTarget,
-          configurationTarget,
-          workspaceFolder.uri
-        );
-        return Logger.infoNotify(
-          `IDF_TARGET has been set to custom. Remember to set the configuration files for OpenOCD`
-        );
-      }
-      await idfConf.writeParameter(
-        "idf.adapterTargetName",
-        selectedTarget.target,
-        configurationTarget,
-        workspaceFolder.uri
-      );
-      if (selectedTarget.target === "esp32") {
-        await idfConf.writeParameter(
-          "idf.openOcdConfigs",
-          ["interface/ftdi/esp32_devkitj_v1.cfg", "target/esp32.cfg"],
-          configurationTarget,
-          workspaceFolder.uri
-        );
-      }
-      if (selectedTarget.target === "esp32s2") {
-        await idfConf.writeParameter(
-          "idf.openOcdConfigs",
-          ["interface/ftdi/esp32_devkitj_v1.cfg", "target/esp32s2.cfg"],
-          configurationTarget,
-          workspaceFolder.uri
-        );
-      }
-      if (
-        selectedTarget.target === "esp32s3" &&
-        selectedTarget.type === "prog"
-      ) {
-        await idfConf.writeParameter(
-          "idf.openOcdConfigs",
-          ["interface/ftdi/esp32_devkitj_v1.cfg", "target/esp32s3.cfg"],
-          configurationTarget,
-          workspaceFolder.uri
-        );
-      }
-      if (
-        selectedTarget.target === "esp32s3" &&
-        selectedTarget.type === "usb"
-      ) {
-        await idfConf.writeParameter(
-          "idf.openOcdConfigs",
-          ["board/esp32s3-builtin.cfg"],
-          configurationTarget,
-          workspaceFolder.uri
-        );
-      }
-      if (
-        selectedTarget.target === "esp32c3" &&
-        selectedTarget.type === "usb"
-      ) {
-        await idfConf.writeParameter(
-          "idf.openOcdConfigs",
-          ["board/esp32c3-builtin.cfg"],
-          configurationTarget,
-          workspaceFolder.uri
-        );
-      }
-      if (
-        selectedTarget.target === "esp32c3" &&
-        selectedTarget.type === "prog"
-      ) {
-        await idfConf.writeParameter(
-          "idf.openOcdConfigs",
-          ["board/esp32c3-ftdi.cfg"],
-          configurationTarget,
-          workspaceFolder.uri
-        );
-      }
-      await vscode.window.withProgress(
-        {
-          cancellable: false,
-          location: vscode.ProgressLocation.Notification,
-          title: "ESP-IDF: Setting device target...",
-        },
-        async (
-          progress: vscode.Progress<{ message: string; increment: number }>
-        ) => {
-          try {
-            if (ConfserverProcess.exists()) {
-              ConfserverProcess.dispose();
-            }
-            const idfPathDir = idfConf.readParameter(
-              "idf.espIdfPath",
-              workspaceFolder.uri
-            );
-            const idfPy = path.join(idfPathDir, "tools", "idf.py");
-            const modifiedEnv = utils.appendIdfAndToolsToPath(
-              workspaceFolder.uri
-            );
-            modifiedEnv.IDF_TARGET = undefined;
-            const enableCCache = idfConf.readParameter(
-              "idf.enableCCache",
-              workspaceFolder.uri
-            ) as boolean;
-            const setTargetArgs: string[] = [idfPy];
-            if (enableCCache) {
-              modifiedEnv.IDF_CCACHE_ENABLE = "1";
-            } else {
-              modifiedEnv.IDF_CCACHE_ENABLE = undefined;
-            }
-            setTargetArgs.push("set-target", selectedTarget.target);
-            const pythonBinPath = idfConf.readParameter(
-              "idf.pythonBinPath",
-              workspaceFolder.uri
-            ) as string;
-            const setTargetResult = await utils.spawn(
-              pythonBinPath,
-              setTargetArgs,
-              {
-                cwd: workspaceFolder.uri.fsPath,
-                env: modifiedEnv,
-              }
-            );
-            Logger.info(setTargetResult.toString());
-            OutputChannel.append(setTargetResult.toString());
-            utils.setCCppPropertiesJsonCompilerPath(workspaceFolder.uri);
-          } catch (err) {
-            if (err.message && err.message.indexOf("are satisfied") > -1) {
-              Logger.info(err.message.toString());
-              OutputChannel.append(err.message.toString());
-            } else {
-              Logger.errorNotify(err, err);
-              OutputChannel.append(err);
-            }
-          }
-        }
-      );
+      await setIdfTarget(enterDeviceTargetMsg);
+      
     });
   });
